requirements.txt
cohere
python-dotenv
flask
requests
